Rename shared folders for clarity
