+++ src/app/routing/route-error-boundary.tsx
@@ -1,6 +1,6 @@
 import Button from "@mui/material/Button";
 import { m } from "@paraglide/messages.js";
-import { ErrorState } from "@shared/components/error-state";
+import { ErrorState } from "@shared/ui/error-state";
 import { useTranslate } from "@shared/language/use-translate";
 import { isRouteErrorResponse, Link, useRouteError } from "react-router";
 import {
