+++ src/features/board/board-sets/board-set-library.tsx
@@ -7,8 +7,8 @@ import ListItemButton from "@mui/material/ListItemButton";
 import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import { m } from "@paraglide/messages.js";
-import { EmptyState } from "@shared/components/empty-state";
-import { LoadingState } from "@shared/components/loading-state";
+import { EmptyState } from "@shared/ui/empty-state";
+import { LoadingState } from "@shared/ui/loading-state";
 import { useTranslate } from "@shared/language/use-translate";
 import { useSnackbar } from "@shared/snackbar/use-snackbar";
 import { useState } from "react";
