+++ src/shared/playback-highlight/highlight-store.test.ts
