+++ src/shared/playback-highlight/highlight-store.ts
