+++ src/shared/ui/error-state.tsx
