+++ src/shared/ui/external-link.test.tsx
