+++ src/shared/ui/external-link.tsx
