+++ src/shared/ui/status-layout.tsx
